requirements.txt
flask>=2.0
tensorflow-cpu>=2.11
Pillow>=9.0
numpy>=1.22
